rustc: split out the pretty-printing parts of ty::print into a separate module.
